Add files via upload
International students, particularly those from China, Thailand, Japan, India, Sri Lanka, etc. prefer to spend less on the living aspect of bachelors or masters study, and by renting a place outside campus, the cost of living reduces by 60% atleast. It’s always chaotic while trying to figure out which product is economical to buy as there are many stores and sources to buy it from. Currently, there is no such system in Syracuse wherein product prices can be compared and bought online through a common marketplace.

With this system in place across all regions in Syracuse, students can make a well-informed choice on their place of residence. Students can also make a checklist of items to buy for the house and order the same, so that it is delivered at their doorstep at the time of arrival (especially in case of incoming students). Students can buy the products from residents – owners or tenants, or from store owners – supermarkets, malls or departmental stores. They are given special student discounts which vary with the product. A student can connect with another fellow student via social media to be updated with the coursework, understand current market situation or even to network and share viewpoints.